+++ Examples/Map+Positioning/Sources/Constants.swift
@@ -1,14 +1,13 @@
 //
 //  Constants.swift
-//  MapExample
+//  Map+PositioningExample
 //
-//  Created by Evgenii Khrushchev on 16/03/2023.
+//  Created by Evgenii Khrushchev on 31/08/2023.
 //  Copyright © 2023 Wemap SAS. All rights reserved.
 //
 
 import Foundation
 import WemapCoreSDK
-import WemapMapSDK
 
 enum Constants {
     
@@ -23,13 +22,6 @@ enum Constants {
     static var polestarApiKey: String {
         fatalError("Specify polestarApiKey and remove fatalError")
     }
-    
-    static var globalNavigationOptions: NavigationOptions {
-        NavigationOptions(
-            itineraryOptions: .init(color: .cyan, projectionOptions: .init(width: 5, color: .lightGray)),
-            userTrackingMode: .followWithHeading
-        )
-    }
 }
 
 enum UIConstants {
@@ -42,12 +34,5 @@ enum UIConstants {
     enum Inset {
         static let top: CGFloat = -200
         static let mid: CGFloat = -150
-    }
-}
-
-extension UserDefaults {
-    
-    static func double(forKey key: PreferencesKey, defaultValue: Double) -> Double {
-        UserDefaults.value(forKey: key) != nil ? standard.double(forKey: key.rawValue) : defaultValue
     }
 }